+++ src/main/java/de/codescape/jira/plugins/scrumpoker/service/ScrumPokerSettingServiceImpl.java
@@ -21,6 +21,7 @@ public class ScrumPokerSettingServiceImpl implements ScrumPokerSettingService {
     private static final String SESSION_TIMEOUT = "sessionTimeout";
     private static final String DEFAULT_PROJECT_ACTIVATION = "defaultProjectActivation";
     private static final String ALLOW_REVEAL_DECK = "allowRevealDeck";
+    private static final String DISPLAY_DROPDOWN_ON_BOARDS = "displayDropdownOnBoards";
 
     private final ActiveObjects activeObjects;
 
@@ -36,6 +37,7 @@ public class ScrumPokerSettingServiceImpl implements ScrumPokerSettingService {
         globalSettings.setSessionTimeout(loadInteger(SESSION_TIMEOUT, SESSION_TIMEOUT_DEFAULT));
         globalSettings.setAllowRevealDeck(AllowRevealDeck.valueOf(loadString(ALLOW_REVEAL_DECK, ALLOW_REVEAL_DECK_DEFAULT.name())));
         globalSettings.setDefaultProjectActivation(loadBoolean(DEFAULT_PROJECT_ACTIVATION, DEFAULT_PROJECT_ACTIVATION_DEFAULT));
+        globalSettings.setDisplayDropdownOnBoards(loadBoolean(DISPLAY_DROPDOWN_ON_BOARDS, DISPLAY_DROPDOWN_ON_BOARDS_DEFAULT));
         return globalSettings;
     }
 
@@ -45,6 +47,7 @@ public class ScrumPokerSettingServiceImpl implements ScrumPokerSettingService {
         persist(STORY_POINT_FIELD, globalSettings.getStoryPointField());
         persist(DEFAULT_PROJECT_ACTIVATION, String.valueOf(globalSettings.isDefaultProjectActivation()));
         persist(ALLOW_REVEAL_DECK, globalSettings.getAllowRevealDeck().name());
+        persist(DISPLAY_DROPDOWN_ON_BOARDS, String.valueOf(globalSettings.isDisplayDropdownOnBoards()));
     }
 
     private String loadString(String key, String defaultValue) {
@@ -58,7 +61,6 @@ public class ScrumPokerSettingServiceImpl implements ScrumPokerSettingService {
         return (scrumPokerSetting != null) ? Integer.valueOf(scrumPokerSetting.getValue()) : defaultValue;
     }
 
-    @SuppressWarnings("SameParameterValue")
     private boolean loadBoolean(String key, boolean defaultValue) {
         ScrumPokerSetting scrumPokerSetting = findByKey(key);
         return (scrumPokerSetting != null) ? Boolean.valueOf(scrumPokerSetting.getValue()) : defaultValue;